Cliente-documentos

+++ avanzo/documento/models.py
@@ -5,8 +5,8 @@ class Documento(models.Model):
     nombre = models.CharField(max_length=100)
     tipo = models.CharField(max_length=20)
     fecha = models.DateField()
-    saldo = models.DecimalField(max_digits=10, decimal_places=2)
-    cuentaBancaria = models.DecimalField(max_digits=15)
+    saldo = models.FloatField()
+    cuentaBancaria = models.FloatField()
     #cliente = models.ForeignKey(Cliente, on_delete=models.CASCADE, default = None)
     def __str__(self):
         return self.nombre
